requirements.txt
numpy==1.19.5
Pillow==8.2.0
tensorflow-cpu==2.5.0
tweepy==3.10.0
pathlib
requests
